+++ src/Workspaces/CSharp/Portable/Formatting/TypingFormattingRule.cs
